requirements.txt
fastapi==0.95.2
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==1.10.7
numpy==1.24.3
pandas==2.0.3
tensorflow==2.13.0
scikit-learn==1.3.0
h5py==3.9.0
pillow==10.0.1
python-multipart==0.0.6
opencv-python-headless==4.8.1.78
tqdm
openpyxl
opencv-python